Move default filter context inside the dashboard staging
JIRA: RAIL-3492
