+++ libs/sdk-ui-dashboard/src/model/_staging/dashboard/defaultFilterContext.ts
